HIA-1420 Move expiry check code
